sports-betting-ai/requirements.txt
streamlit==1.30.0
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
scikit-learn==1.3.2
tensorflow==2.15.0
plotly==5.18.0
python-dotenv==1.0.0
joblib==1.3.2
